v0.99
For mobile screens only. This is not the final version of code.

+++ src/hr/foi/foikviz/NagradeActivity.java
@@ -1,5 +1,7 @@
 package hr.foi.foikviz;
 
+import com.example.foikviz.R;
+
 import android.annotation.SuppressLint;
 import android.app.Activity;
 import android.content.Intent;
@@ -9,17 +11,15 @@ import android.view.View.OnClickListener;
 import android.widget.ImageButton;
 import android.widget.ImageView;
 
-import com.example.foikviz.R;
-
 public class NagradeActivity extends Activity {
 	@Override
 	protected void onCreate(Bundle savedInstanceState) {
 		super.onCreate(savedInstanceState);
 		setContentView(R.layout.activity_nagrade);
-		
-		ImageButton btn_nazad_nagrade = (ImageButton)findViewById(R.id.btn_nazad_nagrade);
+
+		ImageButton btn_nazad_nagrade = (ImageButton) findViewById(R.id.btn_nazad_nagrade);
 		btn_nazad_nagrade.setOnClickListener(new OnClickListener() {
-			
+
 			@Override
 			public void onClick(View v) {
 				Intent i = new Intent(getApplicationContext(),
@@ -30,13 +30,13 @@ public class NagradeActivity extends Activity {
 						R.drawable.pocetna_nagrade_out);
 			}
 		});
-		
+
 		ImageView iv_jojo = (ImageView) findViewById(R.id.jojo);
 		ImageView iv_blok = (ImageView) findViewById(R.id.blok);
 		ImageView iv_knjizica = (ImageView) findViewById(R.id.knjizica);
 		ImageView iv_olovke = (ImageView) findViewById(R.id.olovke);
 		ImageView iv_loptica = (ImageView) findViewById(R.id.loptica);
-		
+
 		float x1 = (float) -3.00;
 		float y1 = 0;
 		float x2 = 0;
@@ -51,32 +51,36 @@ public class NagradeActivity extends Activity {
 		y1 = (float) 0.75;
 		x2 = 0;
 		y2 = 0;
-		PocetniScreenActivity.pokreniAnimaciju(iv_knjizica, x1, y1, x2, y2, 1500);
+		PocetniScreenActivity.pokreniAnimaciju(iv_knjizica, x1, y1, x2, y2,
+				1500);
 		x1 = 3;
 		y1 = (float) 0.75;
 		x2 = 0;
 		y2 = 0;
 		PocetniScreenActivity.pokreniAnimaciju(iv_olovke, x1, y1, x2, y2, 1500);
-		
+
 		x1 = 0;
 		y1 = (float) 5.00;
 		x2 = 0;
 		y2 = 0;
-		PocetniScreenActivity.pokreniAnimaciju(btn_nazad_nagrade, x1, y1, x2, y2, 1500);
-		
+		PocetniScreenActivity.pokreniAnimaciju(btn_nazad_nagrade, x1, y1, x2,
+				y2, 1500);
+
 		x1 = 3;
-		y1 = (float) 0;
+		y1 = 0;
 		x2 = 0;
 		y2 = 0;
-		PocetniScreenActivity.pokreniAnimaciju(iv_loptica, x1, y1, x2, y2, 1500);
-		
+		PocetniScreenActivity
+				.pokreniAnimaciju(iv_loptica, x1, y1, x2, y2, 1500);
+
 	}
-	
+
 	@SuppressLint("NewApi")
 	@Override
 	protected void onResume() {
 		super.onResume();
-		getWindow().getDecorView().setSystemUiVisibility(View.SYSTEM_UI_FLAG_LOW_PROFILE);
+		getWindow().getDecorView().setSystemUiVisibility(
+				View.SYSTEM_UI_FLAG_LOW_PROFILE);
 	}
 
 	@Override
